refactor(insights): eliminate ClickHouse client duplication and fix deprecation warnings
- Create shared ClickHouse client utility module
- Replace deprecated 'host' parameter with 'url' in @clickhouse/client
- Add dynamic protocol detection (HTTP/HTTPS) based on port and env variables
- Simplify database.ts to re-export shared utilities
- Remove ~130 lines of duplicate code from ccusage-utils.tsx
- Update tests to mock shared utilities instead of raw client
- Maintain backward compatibility for all existing APIs
