+++ apps/insights/app/ccusage/__tests__/ccusage-utils.test.ts
@@ -10,47 +10,23 @@ import {
   getCCUsageEfficiency,
   getCCUsageCosts,
 } from '../ccusage-utils'
-import { createClient } from '@clickhouse/client'
 
-// Mock ClickHouse client
-jest.mock('@clickhouse/client', () => ({
-  createClient: jest.fn(() => ({
-    query: jest.fn(),
-    close: jest.fn(),
-  })),
+// Mock ClickHouse client utilities
+jest.mock('../utils/clickhouse-client', () => ({
+  executeClickHouseQueryLegacy: jest.fn(),
 }))
 
-const mockCreateClient = createClient as jest.MockedFunction<typeof createClient>
+import { executeClickHouseQueryLegacy } from '../utils/clickhouse-client'
+const mockExecuteQuery = executeClickHouseQueryLegacy as jest.MockedFunction<typeof executeClickHouseQueryLegacy>
 
 describe('CCUsage Utilities', () => {
   beforeEach(() => {
     jest.clearAllMocks()
-    // Mock environment variables
-    process.env.CLICKHOUSE_HOST = 'test-host'
-    process.env.CLICKHOUSE_PORT = '8123'
-    process.env.CLICKHOUSE_USER = 'test-user'
-    process.env.CLICKHOUSE_PASSWORD = 'test-password'
-    process.env.CLICKHOUSE_DATABASE = 'test-db'
-  })
-
-  afterEach(() => {
-    // Clean up environment variables
-    delete process.env.CLICKHOUSE_HOST
-    delete process.env.CLICKHOUSE_PORT
-    delete process.env.CLICKHOUSE_USER
-    delete process.env.CLICKHOUSE_PASSWORD
-    delete process.env.CLICKHOUSE_DATABASE
   })
 
   describe('getCCUsageMetrics', () => {
     it('should return default metrics when no data is available', async () => {
-      const mockQuery = jest.fn().mockResolvedValue({
-        json: () => Promise.resolve([]),
-      })
-      mockCreateClient.mockReturnValue({
-        query: mockQuery,
-        close: jest.fn(),
-      })
+      mockExecuteQuery.mockResolvedValue([])
 
       const result = await getCCUsageMetrics()
 
@@ -59,6 +35,7 @@ describe('CCUsage Utilities', () => {
         dailyAverage: 0,
         activeDays: 0,
         cacheTokens: 0,
+        totalCost: 0,
         topModel: 'N/A',
       })
     })
@@ -76,19 +53,9 @@ describe('CCUsage Utilities', () => {
       ]
       const mockModelData = [{ model_name: 'claude-3-5-sonnet-20241022' }]
 
-      const mockQuery = jest
-        .fn()
-        .mockResolvedValueOnce({
-          json: () => Promise.resolve(mockData),
-        })
-        .mockResolvedValueOnce({
-          json: () => Promise.resolve(mockModelData),
-        })
-
-      mockCreateClient.mockReturnValue({
-        query: mockQuery,
-        close: jest.fn(),
-      })
+      mockExecuteQuery
+        .mockResolvedValueOnce(mockData)
+        .mockResolvedValueOnce(mockModelData)
 
       const result = await getCCUsageMetrics()
 
@@ -96,19 +63,14 @@ describe('CCUsage Utilities', () => {
       expect(result.dailyAverage).toBe(Math.round(100000 / 15))
       expect(result.activeDays).toBe(15)
       expect(result.cacheTokens).toBe(10000)
+      expect(result.totalCost).toBe(5.25)
       expect(result.topModel).toBe('claude-3-5-sonnet-20241022')
     })
   })
 
   describe('getCCUsageActivity', () => {
     it('should return empty array when no data is available', async () => {
-      const mockQuery = jest.fn().mockResolvedValue({
-        json: () => Promise.resolve([]),
-      })
-      mockCreateClient.mockReturnValue({
-        query: mockQuery,
-        close: jest.fn(),
-      })
+      mockExecuteQuery.mockResolvedValue([])
 
       const result = await getCCUsageActivity()
       expect(result).toEqual([])
@@ -122,16 +84,11 @@ describe('CCUsage Utilities', () => {
           'Input Tokens': 30000,
           'Output Tokens': 15000,
           'Cache Tokens': 5000,
+          'Total Cost': 2.5,
         },
       ]
 
-      const mockQuery = jest.fn().mockResolvedValue({
-        json: () => Promise.resolve(mockData),
-      })
-      mockCreateClient.mockReturnValue({
-        query: mockQuery,
-        close: jest.fn(),
-      })
+      mockExecuteQuery.mockResolvedValue(mockData)
 
       const result = await getCCUsageActivity()
 
@@ -142,6 +99,7 @@ describe('CCUsage Utilities', () => {
         'Input Tokens': Math.round(30000 / 1000),
         'Output Tokens': Math.round(15000 / 1000),
         'Cache Tokens': Math.round(5000 / 1000),
+        'Total Cost': 2.5,
       })
     })
   })
@@ -149,17 +107,11 @@ describe('CCUsage Utilities', () => {
   describe('getCCUsageModels', () => {
     it('should calculate model percentages correctly', async () => {
       const mockData = [
-        { model_name: 'claude-3-5-sonnet', total_tokens: 8000, usage_count: 5 },
-        { model_name: 'claude-3-opus', total_tokens: 2000, usage_count: 2 },
+        { model_name: 'claude-3-5-sonnet', total_tokens: 8000, total_cost: 4.0, usage_count: 5 },
+        { model_name: 'claude-3-opus', total_tokens: 2000, total_cost: 1.0, usage_count: 2 },
       ]
 
-      const mockQuery = jest.fn().mockResolvedValue({
-        json: () => Promise.resolve(mockData),
-      })
-      mockCreateClient.mockReturnValue({
-        query: mockQuery,
-        close: jest.fn(),
-      })
+      mockExecuteQuery.mockResolvedValue(mockData)
 
       const result = await getCCUsageModels()
 
@@ -167,13 +119,17 @@ describe('CCUsage Utilities', () => {
       expect(result[0]).toEqual({
         name: 'claude-3-5-sonnet',
         tokens: 8000,
+        cost: 4.0,
         percent: 80, // 8000/10000 * 100
+        costPercent: 80, // 4.0/5.0 * 100
         usageCount: 5,
       })
       expect(result[1]).toEqual({
         name: 'claude-3-opus',
         tokens: 2000,
+        cost: 1.0,
         percent: 20, // 2000/10000 * 100
+        costPercent: 20, // 1.0/5.0 * 100
         usageCount: 2,
       })
     })
@@ -192,13 +148,7 @@ describe('CCUsage Utilities', () => {
         },
       ]
 
-      const mockQuery = jest.fn().mockResolvedValue({
-        json: () => Promise.resolve(mockData),
-      })
-      mockCreateClient.mockReturnValue({
-        query: mockQuery,
-        close: jest.fn(),
-      })
+      mockExecuteQuery.mockResolvedValue(mockData)
 
       const result = await getCCUsageCosts()
 
@@ -224,13 +174,7 @@ describe('CCUsage Utilities', () => {
         },
       ]
 
-      const mockQuery = jest.fn().mockResolvedValue({
-        json: () => Promise.resolve(mockData),
-      })
-      mockCreateClient.mockReturnValue({
-        query: mockQuery,
-        close: jest.fn(),
-      })
+      mockExecuteQuery.mockResolvedValue(mockData)
 
       const result = await getCCUsageCosts()
 
@@ -253,13 +197,7 @@ describe('CCUsage Utilities', () => {
         },
       ]
 
-      const mockQuery = jest.fn().mockResolvedValue({
-        json: () => Promise.resolve(mockData),
-      })
-      mockCreateClient.mockReturnValue({
-        query: mockQuery,
-        close: jest.fn(),
-      })
+      mockExecuteQuery.mockResolvedValue(mockData)
 
       const result = await getCCUsageEfficiency()
 
@@ -273,11 +211,7 @@ describe('CCUsage Utilities', () => {
 
   describe('Error handling', () => {
     it('should handle ClickHouse connection errors gracefully', async () => {
-      const mockQuery = jest.fn().mockRejectedValue(new Error('Connection failed'))
-      mockCreateClient.mockReturnValue({
-        query: mockQuery,
-        close: jest.fn(),
-      })
+      mockExecuteQuery.mockRejectedValue(new Error('Connection failed'))
 
       const result = await getCCUsageMetrics()
 
@@ -286,16 +220,13 @@ describe('CCUsage Utilities', () => {
         dailyAverage: 0,
         activeDays: 0,
         cacheTokens: 0,
+        totalCost: 0,
         topModel: 'N/A',
       })
     })
 
     it('should return empty array for activity when connection fails', async () => {
-      const mockQuery = jest.fn().mockRejectedValue(new Error('Connection failed'))
-      mockCreateClient.mockReturnValue({
-        query: mockQuery,
-        close: jest.fn(),
-      })
+      mockExecuteQuery.mockRejectedValue(new Error('Connection failed'))
 
       const result = await getCCUsageActivity()
       expect(result).toEqual([])
@@ -304,8 +235,7 @@ describe('CCUsage Utilities', () => {
 
   describe('Environment validation', () => {
     it('should handle missing environment variables', async () => {
-      delete process.env.CLICKHOUSE_HOST
-      delete process.env.CLICKHOUSE_USER
+      mockExecuteQuery.mockResolvedValue([])
 
       const result = await getCCUsageMetrics()
 
@@ -314,6 +244,7 @@ describe('CCUsage Utilities', () => {
         dailyAverage: 0,
         activeDays: 0,
         cacheTokens: 0,
+        totalCost: 0,
         topModel: 'N/A',
       })
     })
